requirements.txt
allure-behave==2.8.40
allure-pytest==2.8.40
allure-python-commons==2.8.40
atomicwrites==1.4.0
attrs==20.3.0
behave==1.2.6
colorama==0.4.4
et-xmlfile==1.1.0
iniconfig==1.1.1
openpyxl==3.0.7
packaging==20.9
parse==1.19.0
parse-type==0.5.2
Pillow==8.2.0
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
selenium==3.141.0
six==1.16.0
toml==0.10.2
unittest-xml-reporting==3.0.4
urllib3==1.26.4
